Hand the qApp object to the main `Kodos` class
Formerly the code used the qApp imported from PyQt4.QtGui which also
seems to work but it is not obviously clear why. It seems like the
qApp there points to the right (the most recent created?) QApplication
object.

Hand a reference to our QApplication object to the `Kodos` class to
make this explicit and more robust.